# Speedup serving configs as json