On my camera the load times for the requests to get_config.sh?conf= changed as follows:
ptz page:
system: before 10s => after 3.6s
ptz_presets: before 3.8s => after 1.7s

mqtt page:
system: before 5.8s => after 1.2s
mqtt: before 4.2s => after 0.7s