@roleoroleo 2025-03-06
Have you tried if this code is affected by sed injection?
I'm not an expert...