@flreinhard 2025-03-06
Short: no it's not.
Long: IMO that's the wrong place for a command injection. It's a static serach/repalce pattern on a external file. Injection could happen if you take a users input (e.g. through website) and pass that as an argument to sed but fail to properly escape the argument/string.