@roleoroleo 2025-03-07
I recently received a mail message about a sed injection on this shell script, that I fixed with this commit:
https://github.com/roleoroleo/sonoff-hack/commit/61f50f5885da0630cb2ec96e16b71d09cea29d0e
If you prepare a special string to use as a content of the post, the risk is a command shell execution in the remote host.
But sed is not my friend and I'm unable to check if your command is affected.