@roleoroleo 2025-03-07
Maybe I should add a validation function to validate the message.